updated routes for passport and package.json
