+++ client/package.json
@@ -7,12 +7,12 @@
   "dependencies": {
     "axios": "^0.18.0",
     "express": "^4.16.4",
-    "mysql": "^2.16.0",
+    "mdbreact": "^4.9.0",
+    "mongoose": "^5.4.4",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.1.1",
-    "sequelize": "^4.42.0"
+    "react-scripts": "^2.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
