+++ client/src/pages/styles/home.css
@@ -1,0 +1,23 @@
+#apppage .gradient {
+    background: rgba(222,128,99,1);
+    background: -moz-linear-gradient(-45deg, rgba(222,128,99,1) 0%, rgba(222,128,99,1) 19%, rgba(217,175,125,1) 91%, rgba(217,175,125,1) 100%);
+    background: -webkit-gradient(left top, right bottom, color-stop(0%, rgba(222,128,99,1)), color-stop(19%, rgba(222,128,99,1)), color-stop(91%, rgba(217,175,125,1)), color-stop(100%, rgba(217,175,125,1)));
+    background: -webkit-linear-gradient(-45deg, rgba(222,128,99,1) 0%, rgba(222,128,99,1) 19%, rgba(217,175,125,1) 91%, rgba(217,175,125,1) 100%);
+    background: -o-linear-gradient(-45deg, rgba(222,128,99,1) 0%, rgba(222,128,99,1) 19%, rgba(217,175,125,1) 91%, rgba(217,175,125,1) 100%);
+    background: -ms-linear-gradient(-45deg, rgba(222,128,99,1) 0%, rgba(222,128,99,1) 19%, rgba(217,175,125,1) 91%, rgba(217,175,125,1) 100%);
+    background: linear-gradient(135deg, rgba(222,128,99,1) 0%, rgba(222,128,99,1) 19%, rgba(217,175,125,1) 91%, rgba(217,175,125,1) 100%);
+    filter: progid:DXImageTransform.Microsoft.gradient( startColorstr='#de8063', endColorstr='#d9af7d', GradientType=1 );
+    
+    }
+    
+    #apppage .view {
+      background-image: url('https://www.pixelstalk.net/cute-baby-animal-wallpapers/#/lightbox&slide=1');
+      background-repeat: no-repeat;
+      background-size: cover;
+      background-position: center center;
+      height: 100vh;
+    }
+    
+    #apppage h6 {
+      line-height: 1.7;
+    }
